Some optimisations

+++ csharp-solution/SpeechFlowCsharp.Tests/AudioProcessing/SpeechSegmenterTests.cs
@@ -15,13 +15,13 @@ namespace SpeechFlowCsharp.Tests.AudioProcessing
             var mockVoiceFilterDetector = new Mock<IVoiceFilter>();
 
             // Configurer le mock pour détecter la parole si la somme des échantillons est positive
-            mockVoiceFilterDetector.Setup(v => v.IsHumanVoice(It.Is<short[]>(s => s.Sum(x => x) > 0))).Returns(true);  // Parole détectée
-            mockVoiceFilterDetector.Setup(v => v.IsHumanVoice(It.Is<short[]>(s => s.Sum(x => x) <= 0))).Returns(false);  // Parole non détectée
+            mockVoiceFilterDetector.Setup(v => v.IsHumanVoice(It.Is<float[]>(s => s.Sum(x => x) > 0))).Returns(true);  // Parole détectée
+            mockVoiceFilterDetector.Setup(v => v.IsHumanVoice(It.Is<float[]>(s => s.Sum(x => x) <= 0))).Returns(false);  // Parole non détectée
 
             var segmenter = new SpeechSegmenter(mockVoiceFilterDetector.Object);
             segmenter.SpeechSegmentDetected += (sender, e) => segmentCount = e.Length;
-            short[] samples = [100, 200, 300];
-            short[] silenceSamples = [0, 0, 0];
+            float[] samples = [100, 200, 300];
+            float[] silenceSamples = [0, 0, 0];
 
             // Act
             // Simuler l'ajout d'échantillons vocaux et la détection de la parole.
