+++ csharp-solution/SpeechFlowCsharp/AudioProcessing/IAudioCapturer.cs
@@ -3,7 +3,7 @@ namespace SpeechFlowCsharp.AudioProcessing
     public interface IAudioCapturer
     {
         bool IsCapturing { get; }
-        event EventHandler<short[]>? AudioCaptured;
+        event EventHandler<float[]>? AudioCaptured;
         void StartCapture();
         void StopCapture();
     }
