+++ csharp-solution/SpeechFlowCsharp/AudioProcessing/IIVadDetector.cs
@@ -3,6 +3,6 @@ namespace SpeechFlowCsharp.AudioProcessing
 {
     public interface IVadDetector
     {
-        bool IsSpeech(short[] samples);
+        bool IsSpeech(float[] samples);
     }
 }
